+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---------------
 - add command-line interface, with export, import, parse, search and stats
 - add option to rename SQL window pages
+- add support for CREATE TABLE .. STRICT (Python 3.9+)
 - update SQLite keywords from 126 to the latest 147-element list
 - use a custom lockfile implementation for single instance check in Linux
 - fix updating main page upon removing databases when home-row is also selected